refactor: apply Suspense to user profile page (#62)
